fixed import in diktat-runner-cli
