+++ diktat-runner/diktat-runner-cli/src/main/kotlin/org/cqfn/diktat/DiktatMain.kt
@@ -4,23 +4,23 @@
 
 package org.cqfn.diktat
 
-import org.cqfn.diktat.api.DiktatError
 import org.cqfn.diktat.api.DiktatMode
+import org.cqfn.diktat.api.DiktatProcessorListener
 import org.cqfn.diktat.cli.DiktatProperties
 import org.cqfn.diktat.common.utils.loggerWithKtlintConfig
-import org.cqfn.diktat.ktlint.unwrap
+import org.cqfn.diktat.ktlint.DiktatProcessorFactoryImpl
 import org.cqfn.diktat.ruleset.utils.isKotlinCodeOrScript
 import org.cqfn.diktat.util.tryToPathIfExists
 import org.cqfn.diktat.util.walkByGlob
 import mu.KotlinLogging
+import java.nio.charset.StandardCharsets
+import java.nio.file.Path
 import java.nio.file.Paths
-import kotlin.io.path.absolutePathString
+import kotlin.io.path.readText
 import kotlin.io.path.writeText
 
 @Suppress("EMPTY_BLOCK_STRUCTURE_ERROR")
 private val log = KotlinLogging.loggerWithKtlintConfig {}
-
-typealias DiktatErrorWithCorrectionInfo = Pair<DiktatError, Boolean>
 
 @Suppress(
     "LongMethod",
@@ -33,17 +33,15 @@ fun main(args: Array<String>) {
     log.debug {
         "Loading diktatRuleSet using config ${properties.config}"
     }
-    val diktatProcessor = DiktatProcessor.builder()
-        .diktatRuleSetProvider(properties.config)
-        .build()
-    val reporter = properties.reporter()
-    reporter.beforeAll()
+    val diktatProcessor = DiktatProcessorFactoryImpl()
+        .create(properties.config)
+    val currentFolder = Paths.get(".")
+    val reporter = properties.reporter(currentFolder)
 
     log.debug {
         "Resolving files by patterns: ${properties.patterns}"
     }
-    val currentFolder = Paths.get(".")
-    properties.patterns
+    val files = properties.patterns
         .asSequence()
         .flatMap { pattern ->
             pattern.tryToPathIfExists()?.let { sequenceOf(it) }
@@ -52,35 +50,26 @@ fun main(args: Array<String>) {
         .filter { file -> file.isKotlinCodeOrScript() }
         .distinct()
         .map { it.normalize() }
-        .map { file ->
+
+    val loggingListener = object : DiktatProcessorListener.Companion.Empty() {
+        override fun before(file: Path) {
             log.debug {
                 "Start processing the file: $file"
             }
-            val result: MutableList<DiktatErrorWithCorrectionInfo> = mutableListOf()
-            DiktatProcessCommand.builder()
-                .processor(diktatProcessor)
-                .file(file)
-                .callback { error, isCorrected ->
-                    result.add(error to isCorrected)
-                }
-                .build()
-                .let { command ->
-                    when (properties.mode) {
-                        DiktatMode.CHECK -> command.check()
-                        DiktatMode.FIX -> {
-                            val formattedFileContent = command.fix()
-                            file.writeText(formattedFileContent, Charsets.UTF_8)
-                        }
-                    }
-                }
-            file to result
         }
-        .forEach { (file, result) ->
-            reporter.before(file.absolutePathString())
-            result.forEach { (error, isCorrected) ->
-                reporter.onLintError(file.absolutePathString(), error.unwrap(), isCorrected)
+    }
+    when (properties.mode) {
+        DiktatMode.CHECK -> diktatProcessor.checkAll(
+            listener = DiktatProcessorListener(loggingListener, reporter),
+            files = files,
+        )
+        DiktatMode.FIX -> diktatProcessor.fixAll(
+            listener = DiktatProcessorListener(loggingListener, reporter),
+            files = files,
+        ) { file, formatterText ->
+            if (file.readText(StandardCharsets.UTF_8) != formatterText) {
+                file.writeText(formatterText, Charsets.UTF_8)
             }
-            reporter.after(file.absolutePathString())
         }
-    reporter.afterAll()
+    }
 }
