+++ diktat-runner/diktat-runner-ktlint-engine/src/main/kotlin/org/cqfn/diktat/ktlint/DiktatBaselineFactoryImpl.kt
@@ -4,6 +4,7 @@ import org.cqfn.diktat.api.DiktatBaseline
 import org.cqfn.diktat.api.DiktatBaselineFactory
 import org.cqfn.diktat.api.DiktatProcessorListener
 import org.cqfn.diktat.api.DiktatProcessorListener.Companion.closeAfterAllAsProcessorListener
+import org.cqfn.diktat.ktlint.DiktatErrorImpl.Companion.wrap
 import org.cqfn.diktat.ktlint.DiktatReporterImpl.Companion.wrap
 import com.pinterest.ktlint.core.internal.loadBaseline
 import com.pinterest.ktlint.reporter.baseline.BaselineReporter
@@ -12,6 +13,10 @@ import java.nio.file.Path
 import kotlin.io.path.absolutePathString
 import kotlin.io.path.outputStream
 
+/**
+ * A factory to create or generate [DiktatBaseline] using `KtLint`
+ */
+@Suppress("DEPRECATION")
 class DiktatBaselineFactoryImpl : DiktatBaselineFactory {
     override fun tryToLoad(baselineFile: Path, sourceRootDir: Path): DiktatBaseline? {
         return loadBaseline(baselineFile.absolutePathString())
